add help text to login/logout commands

+++ htwdresden_bot/login.py
@@ -1,4 +1,3 @@
-import sys
 from telegram.ext import CommandHandler
 from telegram.parsemode import ParseMode
 from htwdresden import RZLogin, Course, HTWAuthenticationException
@@ -6,6 +5,13 @@ from htwdresden_bot import db
 
 
 def _login_cmd(_, update, args):
+    if len(args) == 1 and args[0].lower() == 'hilfe':
+        update.message.reply_text('Hinterlege bei mir deine S-Nummer und dein Passwort, damit ich regelmäßig für dich '
+                                  'nach neuen Noten schauen kann. Wenn ich Änderungen feststellen kann, schicke ich '
+                                  'dir eine Benachrichtigung. Ebenso kannst du nach dem Login den /noten Befehl ohne '
+                                  'Angabe deiner Details nutzen.')
+        return
+
     if len(args) != 2:
         update.message.reply_text('Hierfür benötige ich deine sNummer und dein Passwort. Benutze bitte die Syntax\n'
                                   '`/login s12345 dein_passwort`',
@@ -35,7 +41,12 @@ def _login_cmd(_, update, args):
 login_handler = CommandHandler('login', _login_cmd, pass_args=True)
 
 
-def _logout_cmd(_, update):
+def _logout_cmd(_, update, args):
+    if len(args) == 1 and args[0].lower() == 'hilfe':
+        update.message.reply_text('Mit /logout kannst du deinen bei mir hinterlegten Login löschen. Ich kann dir dann '
+                                  'selbstverständlich auch keine Benachrichtigungen bei neuen Noten mehr schicken.')
+        return
+
     ok = db.remove_login(update.message.chat_id)
     if ok:
         update.message.reply_text('Dein gespeicherter Login wurde erfolgreich gelöscht ✔')
@@ -43,4 +54,4 @@ def _logout_cmd(_, update):
         update.message.reply_text('Dein Login konnte nicht gelöscht werden. Kannte ich ihn überhaupt? 🤔')
 
 
-logout_handler = CommandHandler('logout', _logout_cmd)
+logout_handler = CommandHandler('logout', _logout_cmd, pass_args=True)
